client: move user-related things to their own files
Before this change Users, CreateUser, etc., were in the client.go
file; this splits them out to their own file, which will make it nicer
when adding user removal. It's a separate PR to keep things easily
reviewable.

There is (or should be!) no change to the logic itself; this is meant
to be a pure move.
